+++ src/components/Header/AppBar.jsx
@@ -2,7 +2,6 @@ import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import styles from './style.module.css';
 import { Link } from 'react-router-dom';
-
 import * as ROUTES from '../../constants/routes';
 import * as USER_TYPES from '../../constants/userTypes';
